+++ EstoqueLaravel/database/migrations/2020_12_12_052825_create_saidas_table.php
@@ -18,7 +18,7 @@ class CreateSaidasTable extends Migration
             $table->bigInteger('produto_id')->unsigned()->nullable();
             $table->foreign('produto_id')->references('id')->on('produtos');
             $table->integer('quantidade');
-            $table->decimal('preco_un', 19,2); 
+            $table->double('preco_un', 8, 2); 
             $table->date('data_saida');
             $table->string('observacoes', 4000);
             $table->timestamps();
